+++ source/rafcon/gui/utils/splash_screen.py
@@ -46,9 +46,7 @@ class SplashScreen(Gtk.Window):
         if contains_image:
             main_vbox.pack_start(self.image, True, True, 0)
 
-        if is_custom_design_enabled() and not global_design_config.get_config_value("SPLASH_SCREEN_SHOW_TEXT", True):
-            pass
-        else:
+        if global_design_config.get_config_value("SPLASH_SCREEN_SHOW_TEXT", True):
             # add label to display text, the text can be changed by the text() method.
             # Align it in the middle of the gtk window
             self.label = Gtk.Label(label="")
@@ -56,31 +54,24 @@ class SplashScreen(Gtk.Window):
             self.label.set_yalign(0.5)
             main_vbox.pack_start(self.label, False, True, 10)
             main_vbox.set_spacing(0)
-            label_height = 40
-            if is_custom_design_enabled():
-                label_height = global_design_config.get_config_value("SPLASH_SCREEN_LABEL_HEIGHT")
+            label_height = global_design_config.get_config_value("SPLASH_SCREEN_LABEL_HEIGHT", 0)
             self.label.set_size_request(-1, label_height)
 
         if not os.getenv("RAFCON_START_MINIMIZED", False):
             self.show_all()
 
     def set_text(self, text):
-        if is_custom_design_enabled() and not global_design_config.get_config_value("SPLASH_SCREEN_SHOW_TEXT", True):
+        if not global_design_config.get_config_value("SPLASH_SCREEN_SHOW_TEXT", True):
             return
         logger.info(text)
         self.label.set_text(text)
-        # include this to give more time to watch
-        # import time
-        # time.sleep(1)
         while Gtk.events_pending():
             Gtk.main_iteration_do(False)
         return
 
     def load_image(self, image_path):
         if image_path:
-            horizontal_spacing = 50
-            if is_custom_design_enabled():
-                horizontal_spacing = global_design_config.get_config_value("SPLASH_SCREEN_HORIZONTAL_SPACING")
+            horizontal_spacing = global_design_config.get_config_value("SPLASH_SCREEN_HORIZONTAL_SPACING", 50)
             pixbuf = GdkPixbuf.Pixbuf.new_from_file_at_size(image_path, self.get_size()[0] - horizontal_spacing,
                                                             self.get_size()[1] - horizontal_spacing)
             self.image.set_from_pixbuf(pixbuf)
@@ -91,15 +82,9 @@ class SplashScreen(Gtk.Window):
 
     def get_images(self):
         images = list()
-        if is_custom_design_enabled():
-            splash_screen_path = global_design_config.get_config_value("SPLASH_SCREEN_FOLDER")
-            for image_filename in os.listdir(splash_screen_path):
-                images.append(os.path.join(splash_screen_path, image_filename))
-        else:
-            for image_filename in resource_listdir(rafcon.gui.__name__, os.path.join("assets", "splashscreens")):
-                images.append(resource_filename(rafcon.gui.__name__, os.path.join(
-                    "assets", "splashscreens", image_filename)))
-
+        splash_screen_path = global_design_config.get_config_value("SPLASH_SCREEN_FOLDER")
+        for image_filename in os.listdir(splash_screen_path):
+            images.append(os.path.join(splash_screen_path, image_filename))
         return images
 
     def rotate_image(self, random_=True):