+++ source/rafcon/utils/installation.py
@@ -14,7 +14,7 @@ from os.path import join, dirname, abspath, isfile, isdir
 import sys
 import subprocess
 from distutils.dir_util import copy_tree
-from rafcon.gui.design_config import global_design_config, is_custom_design_enabled
+from rafcon.gui.design_config import global_design_config
 
 from rafcon.utils import resources, log
 
@@ -55,15 +55,14 @@ def install_fonts(restart=False):
     font_names_to_be_installed = ["SourceSansPro", "FontAwesome5Free", "RAFCON"]
 
     custom_fonts = list()  # support a list in case we also support secondary or tertiary fonts
-    if is_custom_design_enabled():
-        custom_fonts.append(global_design_config.get_config_value("PRIMARY_FONT"))
+    custom_fonts.append(global_design_config.get_config_value("PRIMARY_FONT"))
     font_names_to_be_installed += custom_fonts
 
     user_otf_fonts_folder = join(resources.xdg_user_data_folder, "fonts")
 
     font_installed = False
     try:
-        for font_name in font_names_to_be_installed:
+        for font_name in set(font_names_to_be_installed):
             # A font is a folder one or more font faces
             rel_font_folder = join("type1", font_name)
             fonts_folder = resources.get_data_file_path("fonts", rel_font_folder)
@@ -110,12 +109,11 @@ def install_locally_required_files():
         except IOError as e:
             logger.error("Could not copy '{}' files: {}".format(folder, str(e)))
 
-    if is_custom_design_enabled():
-        try:
-            logger.info("Copying custom design files '{}' files...".format(folder))
-            copy_tree(global_design_config.get_config_value("SOURCE_VIEW_FOLDER"),
-                      join(resources.xdg_user_data_folder, "gtksourceview-3.0"), update=1)
-            copy_tree(global_design_config.get_config_value("ICONS_FOLDER"),
-                      join(resources.xdg_user_data_folder, "icons"), update=1)
-        except IOError as e:
-            logger.error("Could not copy '{}' files: {}".format(folder, str(e)))
+    try:
+        logger.info("Copying custom design files '{}' files...".format(folder))
+        copy_tree(global_design_config.get_config_value("SOURCE_VIEW_FOLDER"),
+                  join(resources.xdg_user_data_folder, "gtksourceview-3.0"), update=1)
+        copy_tree(global_design_config.get_config_value("ICONS_FOLDER"),
+                  join(resources.xdg_user_data_folder, "icons"), update=1)
+    except IOError as e:
+        logger.error("Could not copy '{}' files: {}".format(folder, str(e)))