+++ CHANGELOG.rst
@@ -4,10 +4,33 @@ Changelog
 Information about :ref:`RAFCON` changes in each release will be published here. More
 details can be found in the `GIT commit log <https://github.com/DLR-RM/RAFCON/commits/develop>`__.
 
+Future Release
+"""""""
+
 1.1.1
 """""""
 - Bug Fixes:
     - Fix a few GUI bugs
+
+
+1.1.0
+"""""""
+    - Features:
+        - Add skip & skip all buttons in the dialog of the broken libraries during loading a broken session
+        - Set the default directory of the dialog during saving a state machine in accordance with the chosen library in the library tree
+        - Create the data flows & data ports automatically in the nested states
+        - Create the data port automatically when the other state does not have one during connecting the data flows
+        - Support waypoints for data flows
+        - Custom background color for states
+
+    - Bug Fixes:
+        - Fix many minor GUI bugs
+
+
+1.0.1
+"""""""
+    - Bug Fixes:
+        - Fix the default primary font name
 
 
 1.0.0